Implement retrofit and restructure package in core
