+++ core/src/main/java/com/bor3y/core/config/DatabaseConfig.kt
@@ -1,4 +1,4 @@
-package com.bor3y.core
+package com.bor3y.core.config
 
 object DatabaseConfig {
     const val LIST_DB_NAME = "dictations_list.db"
