+++ core/src/main/java/com/bor3y/core/data/local/DictationDatabase.kt
@@ -1,8 +1,8 @@
-package com.bor3y.core.local
+package com.bor3y.core.data.local
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.bor3y.core.local.entity.DictationItemEntity
+import com.bor3y.core.data.local.entity.DictationItemEntity
 
 @Database(
     entities = [DictationItemEntity::class],
