+++ core/src/main/java/com/bor3y/core/data/local/DictationsListDao.kt
@@ -1,10 +1,10 @@
-package com.bor3y.core.local
+package com.bor3y.core.data.local
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Query
 import androidx.room.Upsert
-import com.bor3y.core.local.entity.DictationItemEntity
+import com.bor3y.core.data.local.entity.DictationItemEntity
 
 @Dao
 interface DictationsListDao {
