+++ core/src/main/java/com/bor3y/core/data/remote/DictationsApi.kt
@@ -1,0 +1,11 @@
+package com.bor3y.core.data.remote
+
+import com.bor3y.core.config.ApiConfig
+import com.bor3y.core.data.remote.dto.DictationItemDto
+import retrofit2.http.GET
+
+interface DictationsApi {
+
+    @GET(ApiConfig.LIST_ENDPOINT)
+    suspend fun getDictationsList(): List<Result<DictationItemDto>>
+}
