+++ core/src/main/java/com/bor3y/core/data/remote/dto/DictationItemDto.kt
@@ -1,0 +1,8 @@
+package com.bor3y.core.data.remote.dto
+
+data class DictationItemDto(
+    val createdAt: String,
+    val englishLevel: String,
+    val id: String,
+    val title: String
+)
